rename slog to slogx to avoid conflicts
